Revert "refactor: 代码模块化重构与前端优化"
