+++ config.json
@@ -7,7 +7,7 @@
     "memoryThreshold": 25
   },
   "rotation": {
-    "strategy": "request_count",
+    "strategy": "round_robin",
     "requestCount": 50
   },
   "api": {
@@ -32,7 +32,6 @@
     "retryTimes": 3,
     "skipProjectIdFetch": false,
     "useNativeAxios": false,
-    "useContextSystemPrompt": false,
-    "passSignatureToClient": false
+    "useContextSystemPrompt": false
   }
 }
