+++ public/js/main.js
@@ -7,12 +7,8 @@ initSensitiveInfo();
 // 如果已登录，显示主内容
 if (authToken) {
     showMainContent();
-    restoreTabState(); // 恢复Tab状态
     loadTokens();
-    // 只有在设置页面时才加载配置
-    if (localStorage.getItem('currentTab') === 'settings') {
-        loadConfig();
-    }
+    loadConfig();
 }
 
 // 登录表单提交
