+++ src/utils/imageStorage.js
@@ -1,9 +1,48 @@
 import fs from 'fs';
 import path from 'path';
+import { fileURLToPath } from 'url';
 import config from '../config/config.js';
 import { getDefaultIp } from './utils.js';
-import { getImageDir, isPkg } from './paths.js';
-import { MIME_TO_EXT } from '../constants/index.js';
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = path.dirname(__filename);
+
+// 检测是否在 pkg 打包环境中运行
+const isPkg = typeof process.pkg !== 'undefined';
+
+// 获取图片存储目录
+// pkg 环境下使用可执行文件所在目录或当前工作目录
+function getImageDir() {
+  if (isPkg) {
+    // pkg 环境：优先使用可执行文件旁边的 public/images 目录
+    const exeDir = path.dirname(process.execPath);
+    const exeImageDir = path.join(exeDir, 'public', 'images');
+    try {
+      if (!fs.existsSync(exeImageDir)) {
+        fs.mkdirSync(exeImageDir, { recursive: true });
+      }
+      return exeImageDir;
+    } catch (e) {
+      // 如果无法创建，尝试当前工作目录
+      const cwdImageDir = path.join(process.cwd(), 'public', 'images');
+      try {
+        if (!fs.existsSync(cwdImageDir)) {
+          fs.mkdirSync(cwdImageDir, { recursive: true });
+        }
+        return cwdImageDir;
+      } catch (e2) {
+        // 最后使用用户主目录
+        const homeImageDir = path.join(process.env.HOME || process.env.USERPROFILE || '.', '.antigravity', 'images');
+        if (!fs.existsSync(homeImageDir)) {
+          fs.mkdirSync(homeImageDir, { recursive: true });
+        }
+        return homeImageDir;
+      }
+    }
+  }
+  // 开发环境
+  return path.join(__dirname, '../../public/images');
+}
 
 const IMAGE_DIR = getImageDir();
 
@@ -11,6 +50,14 @@ const IMAGE_DIR = getImageDir();
 if (!isPkg && !fs.existsSync(IMAGE_DIR)) {
   fs.mkdirSync(IMAGE_DIR, { recursive: true });
 }
+
+// MIME 类型到文件扩展名映射
+const MIME_TO_EXT = {
+  'image/jpeg': 'jpg',
+  'image/png': 'png',
+  'image/gif': 'gif',
+  'image/webp': 'webp'
+};
 
 /**
  * 清理超过限制数量的旧图片
